+++ tensorcircuit/backends/numpy_backend.py
@@ -78,6 +78,18 @@ class NumpyBackend(numpy_backend.NumPyBackend, ExtendedBackend):  # type: ignore
         r = np.zeros(shape)
         return self.cast(r, dtype)
 
+    def zeros_like(self, a: Tensor, dtype: Optional[str] = None) -> Tensor:
+        if dtype is None:
+            dtype = self.dtype(a)
+        r = np.zeros_like(a)
+        return self.cast(r, dtype)
+
+    def ones_like(self, a: Tensor, dtype: Optional[str] = None) -> Tensor:
+        if dtype is None:
+            dtype = self.dtype(a)
+        r = np.ones_like(a)
+        return self.cast(r, dtype)
+
     def copy(self, a: Tensor) -> Tensor:
         return a.copy()